5387: move helper function to core
